docs:created components and writePost page
